Move TieJumpPoint code to its own file
